requirements.txt
pytest-html # 测试报告
pymysql
xlrd(1.2.0)